# common-platform-user-mgmt-interfaces/src/main/java/com/firefly/core/users/interfaces/dtos/UserRoleDTO.java
package com.firefly.core.users.interfaces.dtos;

import com.firefly.core.utils.annotations.FilterableId;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.OffsetDateTime;
import java.util.UUID;

/**
 * DTO representing a user-role assignment in the system.
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UserRoleDTO {

    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private UUID id;

    @FilterableId
    private UUID userAccountId;

    @FilterableId
    private UUID roleId;

    @FilterableId
    private UUID branchId;

    @FilterableId
    private UUID distributorId;

    private OffsetDateTime assignedAt;
    private UUID assignedBy;
    private OffsetDateTime createdAt;
    private UUID createdBy;
    private OffsetDateTime updatedAt;
    private UUID updatedBy;
}