# common-platform-user-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-user-mgmt
  description: User Management Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/users
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/roles/{roleId}:
    get:
      tags:
        - Roles
      summary: Get role by ID
      description: Returns a role by its ID
      operationId: getRoleById
      parameters:
        - name: roleId
          in: path
          description: ID of the role to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDTO'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDTO'
    put:
      tags:
        - Roles
      summary: Update an existing role
      description: Updates an existing role and returns the updated role
      operationId: updateRole
      parameters:
        - name: roleId
          in: path
          description: ID of the role to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleDTO'
        required: true
      responses:
        '200':
          description: Role successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDTO'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDTO'
    delete:
      tags:
        - Roles
      summary: Delete a role
      description: Deletes a role by its ID
      operationId: deleteRole
      parameters:
        - name: roleId
          in: path
          description: ID of the role to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role successfully deleted
        '404':
          description: Role not found
  /api/v1/permissions/{permissionId}:
    get:
      tags:
        - Permissions
      summary: Get permission by ID
      description: Returns a permission by its ID
      operationId: getPermissionById
      parameters:
        - name: permissionId
          in: path
          description: ID of the permission to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
        '404':
          description: Permission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
    put:
      tags:
        - Permissions
      summary: Update an existing permission
      description: Updates an existing permission and returns the updated permission
      operationId: updatePermission
      parameters:
        - name: permissionId
          in: path
          description: ID of the permission to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionDTO'
        required: true
      responses:
        '200':
          description: Permission successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
        '404':
          description: Permission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
    delete:
      tags:
        - Permissions
      summary: Delete a permission
      description: Deletes a permission by its ID
      operationId: deletePermission
      parameters:
        - name: permissionId
          in: path
          description: ID of the permission to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Permission successfully deleted
        '404':
          description: Permission not found
...
  /api/v1/users/{userId}/roles:
    get:
      tags:
        - User Roles
      summary: Get roles for a user
      description: Returns all roles assigned to a specific user
      operationId: getRolesByUserId
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved user roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseUserRoleDTO'
    post:
      tags:
        - User Roles
      summary: Assign role to user
      description: Assigns a role to a user and returns the created user role
      operationId: assignRoleToUser
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleDTO'
        required: true
      responses:
        '201':
          description: Role successfully assigned to user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleDTO'
        '404':
          description: User or role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleDTO'
...
  /api/v1/roles/{roleId}/permissions:
    get:
      tags:
        - Role Permissions
      summary: Get permissions for a role
      description: Returns all permissions assigned to a specific role
      operationId: getPermissionsByRoleId
      parameters:
        - name: roleId
          in: path
          description: ID of the role
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved role permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionDTO'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseRolePermissionDTO'
    post:
      tags:
        - Role Permissions
      summary: Assign permission to role
      description: Assigns a permission to a role and returns the created role permission
      operationId: assignPermissionToRole
      parameters:
        - name: roleId
          in: path
          description: ID of the role
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePermissionDTO'
        required: true
      responses:
        '201':
          description: Permission successfully assigned to role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionDTO'
        '404':
          description: Role or permission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionDTO'
...
  /api/v1/roles:
    post:
      tags:
        - Roles
      summary: Create a new role
      description: Creates a new role and returns the created role
      operationId: createRole
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleDTO'
        required: true
      responses:
        '201':
          description: Role successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDTO'
  /api/v1/role-permissions/filter:
    post:
      tags:
        - Role Permissions
      summary: Get all role permissions with filtering
      description: Returns a paginated list of role permissions based on filter criteria
      operationId: filterRolePermissions
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestRolePermissionDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved role permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
  /api/v1/role-permissions:
    post:
      tags:
        - Role Permissions
      summary: Create a new role permission
      description: Creates a new role permission and returns the created role permission
      operationId: createRolePermission
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePermissionDTO'
        required: true
      responses:
        '201':
          description: Role permission successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionDTO'
...
  /api/v1/permissions:
    post:
      tags:
        - Permissions
      summary: Create a new permission
      description: Creates a new permission and returns the created permission
      operationId: createPermission
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionDTO'
        required: true
      responses:
        '201':
          description: Permission successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
...
  /api/v1/role-permissions/{rolePermissionId}:
    get:
      tags:
        - Role Permissions
      summary: Get role permission by ID
      description: Returns a role permission by its ID
      operationId: getRolePermissionById
      parameters:
        - name: rolePermissionId
          in: path
          description: ID of the role permission to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved role permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionDTO'
        '404':
          description: Role permission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionDTO'
    delete:
      tags:
        - Role Permissions
      summary: Delete a role permission
      description: Deletes a role permission by its ID
      operationId: deleteRolePermission
      parameters:
        - name: rolePermissionId
          in: path
          description: ID of the role permission to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role permission successfully deleted
        '404':
          description: Role permission not found
...
  /api/v1/users/{userId}/roles/{roleId}:
    delete:
      tags:
        - User Roles
      summary: Remove role from user
      description: Removes a role from a user
      operationId: removeRoleFromUser
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
            format: uuid
        - name: roleId
          in: path
          description: ID of the role
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role successfully removed from user
        '404':
          description: User role not found
...
  /api/v1/roles/{roleId}/permissions/{permissionId}:
    delete:
      tags:
        - Role Permissions
      summary: Remove permission from role
      description: Removes a permission from a role
      operationId: removePermissionFromRole
      parameters:
        - name: roleId
          in: path
          description: ID of the role
          required: true
          schema:
            type: integer
            format: int64
        - name: permissionId
          in: path
          description: ID of the permission
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Permission successfully removed from role
        '404':
          description: Role permission not found
components:
  schemas:
...
    RoleDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        isAssignable:
          type: boolean
        scopeType:
          type: string
          enum:
            - GLOBAL
            - BRANCH
            - DISTRIBUTOR
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
    PermissionDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        domain:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    UserRoleDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        userAccountId:
          type: string
          format: uuid
        roleId:
          type: string
          format: uuid
        branchId:
          type: string
          format: uuid
        distributorId:
          type: string
          format: uuid
        assignedAt:
          type: string
          format: date-time
        assignedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    RolePermissionDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        roleId:
          type: string
          format: uuid
        permissionId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    FilterRequestRolePermissionDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/RolePermissionDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...
    PaginationResponseUserRoleDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/UserRoleDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    PaginationResponseRolePermissionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/RolePermissionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.